requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
Pillow==10.0.1
numpy==1.24.3
yt-dlp==2023.9.24
requests==2.31.0
matplotlib==3.7.2
scikit-image==0.21.0
google-generativeai==0.3.2
python-dotenv==1.0.0